Fix caching issue whereby changes to CSS are not showing in the browser
New changes to the stylesheet were not being reflected in all browsers.
A hard refresh of the browser fixed the issue. This suggested that
cache-busting was not enabled for this site. A quick inspection proved
this to be true.

As we are using webpack.mix.js we can take advantage of its
cache-busting functionality
(https://laravel.com/docs/6.x/mix#versioning-and-cache-busting).

Using a custom function `mix_asset` in
`web/app/themes/taxcentreofexcellence/lib/setup.php` we can take
advantage of the `mix-manifest.json` file that is automatically
generated when the `dist` assets are built.

The `mix-asset` function was adapted from the `mix-asset` function in
the wp-ppj project
(https://github.com/ministryofjustice/wp-ppj/blob/master/web/app/themes/ppj/functions.php).

Other sites should be checked to see if they have the same issue.